Move MinimalEngine/FillerEngine to engine_wrapper
